requirements.txt
glfw
numpy
pyopengl
pyglm
pyassimp
stbpy
